requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
distro==1.8.0
exceptiongroup==1.0.0
h11==0.14.0
idna==3.4
numpy==1.23.4
outcome==1.2.0
pandas==1.5.1
pdfminer==20191125
pdfminer.six==20220524
pdfplumber==0.7.5
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.15.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==304
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tabula-py==2.5.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
Wand==0.6.10
wsproto==1.2.0
